// src/types/hotel.ts
export interface Hotel {
    id: number;
    nom_fr?: string;
    nom_ar?: string;
    nom_en?: string;
    nom_es?: string;
    nom_pt?: string;
    description_fr?: string;
    description_en?: string;
    description_ar?: string;
    description_es?: string;
    description_pt?: string;
    adresse_fr?: string;
    adresse_en?: string;
    adresse_ar?: string;
    adresse_es?: string;
    adress_pt?: string;
    telephone?: string;
    wilaya_fr?: string;
    wilaya_ar?: string;
    wilaya_en?: string;
    wilaya_es?: string;
    wilaya_pt?: string;
    email?: string;
    site_web?: string;
    etoiles?: number;
    prix_min?: number;
    prix_max?: number;
    latitude?: number;
    longitude?: number;
    image_url?: string;
    services_fr?: string;
    services_en?: string;
    services_ar?: string;
    services_es?: string;
    services_pt?: string;
    statut?: string;
    debut_depart?: Date;
    fin_depart?: Date;
    intervalle_depart?: number;
    debut_retour?: Date;
    fin_retour?: Date;
    intervalle_retour?: number;
    zone_fr?: string;
    zone_ar?: string;
    zone_en?: string;
    zone_es?: string;
    zone_pt?: string;
    quai_fr?: string;
    quai_ar?: string;
    quai_en?: string;
    quai_es?: string;
    quai_pt?: string;
    distance?: number;
    tempestimer?: Date;
    date_creation?: Date;
    date_modification?: Date;
  }
  
  export type ViewMode = 'card' | 'list';